dev modify

+++ src/main/java/com/lorren/dao/AccountDAO.java
@@ -4,6 +4,6 @@ import com.lorren.entity.Account;
 
 public interface AccountDAO {
 
-    long insert(Account account);
+    int insert(Account account);
 
 }
